This is the second patch for RC6.
This patch includes all the fixes from the first patch, so it can be the only patch you need. Until otherwise notified, the newest patch is the only one you need. The patching process today only updates the source code. When we change that, then we may need to force some order to patches, and that will be expressed in the notes. 

The fixes in this patch expand the touch area around the chevron of the tree view, so on a phone, your experience will be that you expand the tree as opposed to clicking to the item directly. It's a more intuitive interface I believe.