Add files via upload
With this new version update, it might be recommended to do a backup before you install this. It doesn't depend upon an NGINX configuration change, but it's only been tested with an NGINX configuration change.The patch doesn't have yet the ability to modify the Nginx configuration.The configuration in question is simply some caching, so I believe this should be a safe update. But back up just in case.